Order updates multiple shipping addresses (#1742)
* Order updates multiple shipping addresses

* Review remarks inclusion

+++ commercetools-models/src/main/java/io/sphere/sdk/carts/ItemShippingDetails.java
@@ -23,7 +23,7 @@ public interface ItemShippingDetails {
      * Convenience method to extract map address keys to their target without further processing
      * @return Map with addressKey as key and destined quantity as value.
      */
-    default Map<String, Integer> getTargetsMap() {
+    default Map<String, Long> getTargetsMap() {
         return getTargets().stream().collect(Collectors.toMap(ItemShippingTarget::getAddressKey, ItemShippingTarget::getQuantity));
     }
