+++ commercetools-models/src/main/java/io/sphere/sdk/carts/ItemShippingTarget.java
@@ -12,6 +12,6 @@ public interface ItemShippingTarget {
 
     String getAddressKey();
 
-    int getQuantity();
+    Long getQuantity();
 
 }
